# Conflicts:
#	src/app/services/admin.service.ts

+++ src/app/services/appointment.service.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from "@angular/common/http";
+import { ParseArgumentException } from "@angular/cli/models/parser";
 import { Observable } from "rxjs";
 import { environment } from "../../environments/environment";
 
@@ -7,6 +8,7 @@ import { Appointment } from "../types/appointment";
 import { TimespanId } from "../types/aliases/timespan-id";
 import { RoomId } from "../types/aliases/room-id";
 import { ConfirmationStatus } from "../types/enums/confirmation-status";
+import { SeriesId } from "../types/aliases/series-id";
 
 @Injectable({
   providedIn: 'root'
@@ -14,6 +16,8 @@ import { ConfirmationStatus } from "../types/enums/confirmation-status";
 
 /**
  * Service for the management of appointments
+ * @typedef {Service} AppointmentService
+ * @class
  */
 export class AppointmentService {
 
@@ -24,7 +28,7 @@ export class AppointmentService {
    * Retrieves all appointments
    */
   public getAllAppointments(): Observable<any> {
-    const apiURL = `${environment.baseUrl}${environment.apiRoutes.allAppointments}`;
+    const apiURL = `${environment.baseUrl}${environment.apiRoutes.appointments.getAllAppointments}`;
 
     return this.httpClient.get(apiURL);
   }
@@ -33,7 +37,7 @@ export class AppointmentService {
    * Retrieves all appointments for current user
    */
   public getAllAppointmentsForCurrentUser(): Observable<any> {
-    const apiURL = `${environment.baseUrl}${environment.apiRoutes.myAppointments}`;
+    const apiURL = `${environment.baseUrl}${environment.apiRoutes.appointments.getCurrentUserAppointments}`;
 
     return this.httpClient.get(apiURL);
   }
@@ -44,7 +48,11 @@ export class AppointmentService {
    * @param {RoomId} roomId id of room to retrieve appointments
    */
   public getAllAppointmentsForRoom(roomId: RoomId): Observable<any> {
-    const apiURL = `${environment.baseUrl}${environment.apiRoutes.appointmentsForRoom
+    if (roomId === null) {
+      throw ParseArgumentException;
+    }
+
+    const apiURL = `${environment.baseUrl}${environment.apiRoutes.appointments.getRoomAppointments
       .replace(':id', roomId)}`;
 
     return this.httpClient.get(apiURL);
@@ -56,8 +64,28 @@ export class AppointmentService {
    * @param {TimespanId} appointmentId id of the appointment
    */
   public getAppointmentData(appointmentId : TimespanId): Observable<any> {
-    const apiURL = `${environment.baseUrl}${environment.apiRoutes.viewAppointment
+    if (appointmentId === null) {
+      throw ParseArgumentException;
+    }
+
+    const apiURL = `${environment.baseUrl}${environment.apiRoutes.appointments.getSingleAppointment
       .replace(':id', appointmentId)}`;
+
+    return this.httpClient.get(apiURL);
+  }
+
+  /**
+   * Retrieves all data of the appointments for one series
+   *
+   * @param {SeriesId} seriesId id of the appointment
+   */
+  public getAllAppointmentsForSeries(seriesId : SeriesId): Observable<any> {
+    if (seriesId === null) {
+      throw ParseArgumentException;
+    }
+
+    const apiURL = `${environment.baseUrl}${environment.apiRoutes.appointments.getSeriesAppointments
+      .replace(':id', seriesId)}`;
 
     return this.httpClient.get(apiURL);
   }
@@ -68,9 +96,27 @@ export class AppointmentService {
    * @param {Appointment} appointment all data about the requested appointment
    */
   public createAppointment(appointment: Appointment): Observable<any> {
-    const apiURL = `${environment.baseUrl}${environment.apiRoutes.createAppointment}`;
+    const apiURL = `${environment.baseUrl}${environment.apiRoutes.appointments.createAppointment}`;
     const requestBody = {
       appointment: appointment,
+    };
+
+    return this.httpClient.post(apiURL, {headers: requestBody});
+  }
+
+  /**
+   * Creates a new appointment request for a series of appointments
+   *
+   * @param {Appointment} appointment the starting appointment of the series
+   * @param {number} difference milliseconds, time difference between the appointments, regularity
+   * @param {number} amount 2-2048, amount of appointments wanted for the series
+   */
+  public createAppointmentSeries(appointment: Appointment, difference: number, amount: number) {
+    const apiURL = `${environment.baseUrl}${environment.apiRoutes.appointments.createAppointmentSeries}`;
+    const requestBody = {
+      appointment: appointment,
+      difference: difference,
+      amount: amount
     };
 
     return this.httpClient.post(apiURL, {headers: requestBody});
@@ -82,11 +128,34 @@ export class AppointmentService {
    * @param {TimespanId} appointmentId id of the appointment to be edited
    * @param {object} changedData   changed values as object
    */
-  public editAppointment(appointmentId : TimespanId, changedData: object): Observable<any> {
-    const apiURL = `${environment.baseUrl}${environment.apiRoutes.editAppointment
+  public editAppointment(appointmentId: TimespanId, changedData: object): Observable<any> {
+    if (appointmentId === null) {
+      throw ParseArgumentException;
+    }
+    const apiURL = `${environment.baseUrl}${environment.apiRoutes.appointments.updateAppointment
       .replace(':id', appointmentId)}`;
     const requestBody = {
       appointmentId: appointmentId,
+      changedData: changedData
+    };
+
+    return this.httpClient.patch(apiURL, {headers: requestBody});
+  }
+
+  /**
+   * Edits a series of appointment
+   *
+   * @param {SeriesId} seriesId id of a series of appointment to be edited
+   * @param {object} changedData   changed values as object
+   */
+  public editAppointmentSeries(seriesId: SeriesId, changedData: object): Observable<any> {
+    if (seriesId === null) {
+      throw ParseArgumentException;
+    }
+    const apiURL = `${environment.baseUrl}${environment.apiRoutes.appointments.updateAppointmentSeries
+      .replace(':id', seriesId)}`;
+    const requestBody = {
+      appointmentId: seriesId,
       changedData: changedData
     };
 
@@ -99,8 +168,28 @@ export class AppointmentService {
    * @param {TimespanId} appointmentId Id of an appointment
    */
   public deleteAppointment(appointmentId: TimespanId): Observable<any> {
-    const apiURL = `${environment.baseUrl}${environment.apiRoutes.deleteAppointment
+    if (appointmentId === null) {
+      throw ParseArgumentException;
+    }
+
+    const apiURL = `${environment.baseUrl}${environment.apiRoutes.appointments.deleteAppointment
       .replace(':id', appointmentId)}`;
+
+    return this.httpClient.delete(apiURL);
+  }
+
+  /**
+   * Deletes a series of appointment
+   *
+   * @param {SeriesId} seriesId Id of an appointment
+   */
+  public deleteAppointmentSeries(seriesId: SeriesId): Observable<any> {
+    if (seriesId === null) {
+      throw ParseArgumentException;
+    }
+
+    const apiURL = `${environment.baseUrl}${environment.apiRoutes.appointments.deleteAppointmentSeries
+      .replace(':id', seriesId)}`;
 
     return this.httpClient.delete(apiURL);
   }
@@ -115,11 +204,29 @@ export class AppointmentService {
   }
 
   /**
-   * Sets appointment request to accepted
+   * Sets appointment request to declined
    *
    * @param {TimespanId} appointmentId id of appointment
    */
   public declineAppointmentRequest(appointmentId: TimespanId): Observable<any> {
     return this.editAppointment(appointmentId, { confirmationStatus: ConfirmationStatus.denied });
+  }
+
+  /**
+   * Sets series appointments request to accepted
+   *
+   * @param {SeriesId} seriesId id of series of appointments
+   */
+  public acceptAppointmentSeriesRequest(seriesId: SeriesId): Observable<any> {
+    return this.editAppointmentSeries(seriesId, { confirmationStatus: ConfirmationStatus.denied });
+  }
+
+  /**
+   * Sets series of appointments request to declined
+   *
+   * @param {SeriesId} seriesId id of series of appointments
+   */
+  public declineAppointmentSeriesRequest(seriesId: SeriesId): Observable<any> {
+    return this.editAppointmentSeries(seriesId, { confirmationStatus: ConfirmationStatus.denied });
   }
 }
